model [nfc]: Pull out new api/model/reaction.dart
We'll add more things here soon.
